+++ client/src/components/common/LoadingIndicator.js
