+++ client/src/components/pages/Main/index.js
@@ -1,16 +1,16 @@
 import React, { useRef, useState } from 'react'
 import styled, { css } from 'styled-components'
-import Tab from './Tab'
-import API from '../util/api'
-import TabPanel from './TabPanel'
-import Container from './Container'
-import Badge from './Badge'
-import Button from './Button'
-import Counter from './Counter'
-import Payment from './Payment'
-import LoadingIndicator from './LoadingIndicator'
-import Receipt from './Receipt'
-import Input from './Input'
+import Tab from '../../layout/Tab'
+import API from '../../../util/api'
+import TabPanel from '../../layout/TabPanel'
+import Container from '../../common/Container'
+import Badge from '../../common/Badge'
+import Button from '../../common/Button'
+import Counter from '../../common/Counter'
+import Payment from '../../layout/Payment'
+import LoadingIndicator from '../../common/LoadingIndicator'
+import Receipt from '../../layout/Receipt'
+import Input from '../../common/Input'
 
 const ItemsLayout = styled.section`
   display: flex;
